(*) New DTO

+++ src/main/java/agp/sanclemen/agp_proyecto/model/Customer.java
@@ -11,7 +11,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID", nullable = false)
-    private long id;
+    private Long id;
 
     @Column(name = "NAME", nullable = false)
     private String name;
@@ -47,7 +47,7 @@ public class Customer {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
